add command-line parameter to remove data

+++ clean.sh
@@ -1,4 +1,7 @@
 #!/bin/bash
 
 rm log/*f9beb4d9*
-rm data/crawl/f9beb4d9/*.json data/crawl/f9beb4d9/*.csv
+
+if [ "$1" == "--hard" ]; then 
+    rm data/crawl/f9beb4d9/*.json data/crawl/f9beb4d9/*.csv
+fi
